Further organisational improvement of store functions code, focusing on store get code.
