+++ src/store/get/queryPlan/dataNodes.spec.ts
@@ -1,5 +1,5 @@
 import { Operator } from '@samhuk/data-filter/dist/types'
-import { tsPgOrm } from '../../testData'
+import { tsPgOrm } from '../../../testData'
 import { toDataNodes } from './dataNodes'
 
 describe('dataNodes', () => {
