+++ src/store/get/queryPlan/queryPlan.ts
@@ -1,9 +1,9 @@
 import { SimplePgClient } from 'simple-pg-client/dist/types'
-import { DataFormat, DataFormatDeclarations, DataFormatsDict } from '../../dataFormat/types'
-import { removeDuplicates, removeNullAndUndefinedValues } from '../../helpers/array'
-import { deepRemovePropsWithPrefix } from '../../helpers/object'
-import { RelationDeclarations, RelationsDict, RelationType } from '../../relations/types'
-import { GetFunctionOptions } from '../types/get'
+import { DataFormat, DataFormatDeclarations, DataFormatsDict } from '../../../dataFormat/types'
+import { removeDuplicates, removeNullAndUndefinedValues } from '../../../helpers/array'
+import { deepRemovePropsWithPrefix } from '../../../helpers/object'
+import { RelationDeclarations, RelationsDict, RelationType } from '../../../relations/types'
+import { GetFunctionOptions } from '../types'
 import { toDataNodes } from './dataNodes'
 import { toQueryNodes } from './queryNodes'
 import { DataNode, QueryNode, QueryNodes, QueryNodeSql, QueryPlan } from './types'
