+++ src/store/types/index.ts
@@ -1,7 +1,7 @@
 import { DataFormatDeclarations } from '../../dataFormat/types'
 import { RelationDeclarations } from '../../relations/types'
 import { DeleteSingleFunction } from '../delete/types'
-import { GetSingleFunction, GetMultipleFunction } from './get'
+import { GetSingleFunction, GetMultipleFunction } from '../get/types'
 import { UpdateSingleFunction } from '../update/types'
 import { CreateManualSingleFunction, CreateSingleFunction } from '../create/types'
